Move test to test_wal_proposer.py
